+++ backend/src/main/java/com/meterease/backend/repository/ManagerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ManagerRepository extends JpaRepository<Manager, Long> {
+public interface ManagerRepository extends JpaRepository<Manager, Integer> {
 
     Optional<Manager> findByEmailAddress(String emailAddress);
